fix(coaching): resolve Phase 1 MINOR review issues
1. MSRV: replace is_multiple_of(3) with % 3 == 0 (needs 1.83+, MSRV is 1.77.1)
2. GoalProgress percentage: u8 -> u16 to support >100% goal completion
3. Placeholder "N/A" variables: add TODO(Phase 2) comments with wiring targets
4. regime_duration_secs = poll interval: document overstay trigger inertness
5. Implicit feedback (None, ""): document need for real regime/app data

+++ crates/oneshim-core/src/models/coaching.rs
@@ -105,8 +105,8 @@ pub struct GoalProgress {
     pub current_minutes: u32,
     /// User-configured daily target for this regime.
     pub target_minutes: u32,
-    /// Completion percentage (0-100+, can exceed 100%).
-    pub percentage: u8,
+    /// Completion percentage. Values >100 are valid (user exceeded the goal).
+    pub percentage: u16,
 }
 
 /// Extended goal progress with UI display hints.
@@ -118,8 +118,8 @@ pub struct GoalProgressView {
     pub current_minutes: u32,
     /// User-configured daily target.
     pub target_minutes: u32,
-    /// Completion percentage.
-    pub percentage: u8,
+    /// Completion percentage. Values >100 are valid (user exceeded the goal).
+    pub percentage: u16,
     /// CSS-compatible color string for progress bar rendering.
     pub display_color: String,
 }
